+++ tests/integration/cases/call_ref_args_trivial_register/Bash_call.sh
@@ -1,0 +1,13 @@
+process() { :; }
+declare my_int=1
+declare my_bool=true
+declare my_float=3.14
+declare my_list=(
+    1
+    2
+    3
+)
+process my_int 42
+process my_bool 7
+process my_float 9
+process my_list 1